Renamed resource and added to project readme
